cv-synth-backend/requirements.txt
annotated-types==0.7.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.0.4
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
Flask==2.2.2
Flask-Cors==3.0.10
google-ai-generativelanguage==0.2.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.5.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.62.3
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
MarkupSafe==3.0.2
murmurhash==1.0.13
numpy==2.2.6
packaging==25.0
pathlib_abc==0.1.1
pathy==0.11.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==0.21.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
smart-open==6.4.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.4.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==0.10.1
weasel==0.4.1
Werkzeug==2.3.8
# Added for file parsing
python-docx==1.1.0
pdfplumber==0.10.3